GeoHospital/src/requirements.txt
Django==3.1.4
djangorestframework==3.12.2
asgiref==3.3.1
gdal==2.4.0
postgis==1.0.4
psycopg2-binary==2.8.6
pytz==2021.1
sqlparse==0.4.1
django-filter==2.4.0